update simulator, add new map, add other players

+++ src/cell/sim/Cell.java
@@ -9,11 +9,11 @@ public class Cell {
 	// configuration info that varies less
 	private static boolean gui = false;
 	private static int turns = 100;
-	private static int traders = 10;
-	private static int marbles = 10;
-	private static int dim = 5;
-	private static String mapPath = "cell/map/g3-traps.txt";
-	private static String playerPath = "cell/players.list";
+	private static int traders = 2;
+	private static int marbles = 20;
+	private static boolean recompile = true;
+	private static String mapPath = "g5-single.txt";
+	private static String playerPath = "players.list";
 
 	// return game turns
 	public static int gameTurns() { return turns; }
@@ -38,46 +38,55 @@ public class Cell {
 	}
 
 	// compile and load players dynamically
-	private static Player[] loadPlayers(String txtPath) {
+	private static Player[] loadPlayers(String txtPath) throws Exception {
 		// list of players
 		List <Player> playersList = new LinkedList <Player> ();
-		try {
-			// get file of players
-			BufferedReader in = new BufferedReader(new FileReader(new File(txtPath)));
-			// get tools
-			ClassLoader loader = ToolProvider.getSystemToolClassLoader();
-			JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
-			StandardJavaFileManager fileManager = compiler.getStandardFileManager(null, null, null);
-			// load players
-			String group;
-			while ((group = in.readLine()) != null) {
-				System.err.println("Group: " + group);
+		// get file of players
+		BufferedReader in = new BufferedReader(new FileReader(new File(txtPath)));
+		// get tools
+		ClassLoader loader = Cell.class.getClassLoader();
+		if (loader == null) throw new Exception("Cannot load class loader");
+		JavaCompiler compiler = null;
+		StandardJavaFileManager fileManager = null;
+		// get separator
+		String sep = File.separator;
+		// load players
+		String group;
+		while ((group = in.readLine()) != null) {
+			System.err.println("Group: " + group);
+			// search for compiled files
+			File classFile = new File("cell" + sep + group + sep + "Player.class");
+			System.err.println(classFile.getAbsolutePath());
+			if (!classFile.exists() || recompile) {
 				// delete all class files
-				List <File> classFiles = directoryFiles("src/cell/" + group, ".class");
+				List <File> classFiles = directoryFiles("src/cell" + sep + group, ".class");
 				System.err.print("Deleting " + classFiles.size() + " class files...   ");
 				for (File file : classFiles)
 					file.delete();
 				System.err.println("OK");
+				if (compiler == null) compiler = ToolProvider.getSystemJavaCompiler();
+				if (compiler == null) throw new Exception("Cannot load compiler");
+				if (fileManager == null) fileManager = compiler.getStandardFileManager(null, null, null);
+				if (fileManager == null) throw new Exception("Cannot load file manager");
 				// compile all files
-				List <File> javaFiles = directoryFiles("src/cell/" + group, ".java");
+				List <File> javaFiles = directoryFiles("src/cell" + sep + group, ".java");
 				System.err.print("Compiling " + javaFiles.size() + " source files...   ");
 				Iterable<? extends JavaFileObject> units = fileManager.getJavaFileObjectsFromFiles(javaFiles);
 				boolean ok = compiler.getTask(null, fileManager, null, null, null, units).call();
-				if (!ok) throw new Exception("compile error");
+				if (!ok) throw new Exception("Compile error");
 				System.err.println("OK");
-				// load class
-				System.err.print("Loading player class...   ");
-				Class playerClass = loader.loadClass("cell." + group + ".Player");
-				System.err.println("OK");
-				// set name of player and append on list
-				Player player = (Player) playerClass.newInstance();
-				playersList.add(player);
 			}
-			in.close();
-		} catch (Exception e) {
-			e.printStackTrace(System.err);
-			return null;
+			// load class
+			System.err.print("Loading player class...   ");
+			Class playerClass = loader.loadClass("cell." + group + ".Player");
+			System.err.println("OK");
+			// set name of player and append on list
+			Player player = (Player) playerClass.newInstance();
+			if (player == null)
+				throw new Exception("Load error");
+			playersList.add(player);
 		}
+		in.close();
 		return playersList.toArray(new Player[0]);
 	}
 
@@ -140,24 +149,26 @@ public class Cell {
 
 	public static void main(String[] args) throws Exception
 	{
-		// board size
-		if (args.length > 1)
-			dim = Integer.parseInt(args[1]);
 		// starting marbles
-		if (args.length > 2)
-			marbles = Integer.parseInt(args[2]);
+		if (args.length > 1)
+			marbles = Integer.parseInt(args[1]);
 		// traders
-		if (args.length > 3)
-			traders = Integer.parseInt(args[3]);
+		if (args.length > 2)
+			traders = Integer.parseInt(args[2]);
 		// game turns
-		if (args.length > 4)
-			turns = Integer.parseInt(args[4]);
+		if (args.length > 3)
+			turns = Integer.parseInt(args[3]);
 		// map path
-		if (args.length > 5)
-			mapPath = args[5];
+		String sep = File.separator;
+		if (args.length > 4)
+			mapPath = args[4];
+		else
+			mapPath = "cell" + sep + "map" + sep + mapPath;
 		// players path
-		if (args.length > 6)
-			playerPath = args[6];
+		if (args.length > 5)
+			playerPath = args[5];
+		else
+			playerPath = "cell" + sep + playerPath;
 		Cell game = new Cell();
 		game.printPlayers();
 		game.printTraders();
@@ -166,12 +177,18 @@ public class Cell {
 		char req = 'X';
 		if (gui) {
 			server = new HTTPServer();
+			int port = server.port();
+			System.err.println("Port: " + port);
+			System.out.println("Port: " + port);
 			while ((req = server.nextRequest(0)) == 'I');
 			if (req != 'B')
 				throw new Exception("Invalid first request");
 		}
 		for (File f : directoryFiles("webpages", ".html"))
 			f.delete();
+		FileOutputStream out = new FileOutputStream("webpages/index.html");
+		out.write(game.state().getBytes());
+		out.close();
 		for (int t = 1 ; t <= turns; ++t) {
 			boolean f = true;
 			if (server != null) do {
@@ -183,7 +200,7 @@ public class Cell {
 				f = false;
 			} while (req == 'B');
 			boolean end = game.next();
-			FileOutputStream out = new FileOutputStream("webpages/" + t + ".html");
+			out = new FileOutputStream("webpages/" + t + ".html");
 			out.write(game.state().getBytes());
 			out.close();
 			game.printPlayers();
@@ -191,11 +208,10 @@ public class Cell {
 		}
 		game.rank();
 		if (server != null) {
-			server.replyState(game.state(), refresh);
+			server.replyState(game.state(), 0);
 			while ((req = server.nextRequest(2000)) == 'I');
-			if (req != 'X')
-				throw new Exception("Invalid last request");
 		}
+		server.close();
 	}
 
 	private Player[] players;
@@ -208,6 +224,7 @@ public class Cell {
 	private int[] round;
 	private int[] count;
 	private boolean[] in;
+	private int dim;
 	private int turn;
 	private int trades;
 	private int conflicts;
@@ -217,30 +234,84 @@ public class Cell {
 	private String state()
 	{
 		int pixels = 800;
-		String title = "PPS Cell";
+		String title = "CELL";
 		StringBuffer buf = new StringBuffer("");
 		buf.append("<!DOCTYPE html PUBLIC \"-//W3C//DTD XHTML 1.0 Transitional//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\">\n");
 		buf.append("<html xmlns=\"http://www.w3.org/1999/xhtml\" dir=\"ltr\" lang=\"en-US\" xml:lang=\"en\">\n");
 		buf.append("<head>\n");
+		buf.append(" <meta http-equiv=\"Content-Type\" content=\"text/html; charset=iso-8859-7\" />\n");
+		buf.append(" <link rel=\"shortcut icon\" href=\"cell/icon.ico\" />\n");
 		buf.append(" <title>" + title + "</title>\n");
-		buf.append(" <style>a:hover {color:red;}</style>\n");
+		buf.append(" <style type=\"text/css\">\n");
+		buf.append("  a:link {text-decoration: none; color: blue;}\n");
+		buf.append("  a:visited {text-decoration: none; color: blue;}\n");
+		buf.append("  a:hover {text-decoration: none; color: red;}\n");
+		buf.append("  a:active {text-decoration: none; color: blue;}\n");
+		buf.append(" </style>\n");
 		buf.append("</head>\n");
 		buf.append("<body>\n");
-		buf.append(" <div style=\"width:" + (pixels + 550) + "px; margin-left:auto; margin-right: auto;\">\n");
-		buf.append("  <div style=\"width: 200px; float: left;\">\n");
-		buf.append("   <div style=\"width: 200px; height: 120px;\"></div>\n");
+		// general part
+		buf.append(" <div style=\"width:" + (pixels + 400) + "px; margin-left:auto; margin-right: auto;\">\n");
+		// left part
+		buf.append("  <div style=\"width: 400px; float: left;\">\n");
+		// space above
+		buf.append("   <div style=\"width: 400px; height: 25px;\"></div>\n");
 		buf.append("   <div style=\"clear:both;\"></div>\n");
-		buf.append("   <a href=\"play\"><div style=\"width: 200px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;");
-		buf.append("                              color: black; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">Play</div></a>\n");
-		buf.append("   <div style=\"clear:both;\"></div>\n");
-		buf.append("   <a href=\"stop\"><div style=\"width: 200px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;");
-		buf.append("                              color: black; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">Stop</div></a>\n");
-		buf.append("   <div style=\"clear:both;\"></div>\n");
-		buf.append("   <a href=\"step\"><div style=\"width: 200px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;");
-		buf.append("                              color: black; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">Step</div></a>\n");
-		buf.append("  </div>\n");
-		buf.append("  <div style=\"width:" + pixels + "px; float: left;\">\n");
+		// array of results below buttons
+		buf.append("   <div style=\"width: 400px; float: left;\">\n");
+		// empty up left square
+		buf.append("    <div style=\"width: 40px; height: 50px; float:left;\"></div>\n");
+		// color squares
 		String[] colors = {"red", "green", "blue", "yellow", "purple", "orange"};
+		for (String color : colors)
+			buf.append("    <div style=\"width: 44px; height: 44px; float:left; border: 3px solid black; background-color: " + color + "\"></div>\n");
+		// empty up right square
+		buf.append("    <div style=\"width: 60px; height: 50px; float:left;\"></div>\n");
+		buf.append("    <div style=\"clear:both;\"></div>\n");
+		// result lines
+		for (int p = 0 ; p != players.length ; ++p) {
+			// player name
+			buf.append("    <div style=\"width: 34px; height: 44px; float:left; border: 3px solid black; text-align: center;");
+			if (in[p]) buf.append("\n");
+			else buf.append(" text-decoration: line-through;\n");
+			buf.append("                font-size: 25px; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">" + (p + 1) + "</div>\n");
+			int total = 0;
+			// marble score per color
+			buf.append("    <div style=\"width: 294px; height: 44px; float: left; border: 3px solid black;\">\n");
+			for (int r = 0 ; r != 6 ; ++r) {
+				String s = "";
+				if (in[p]) s += sacks[p][r];
+				buf.append("     <div style=\"width: 49px; height: 44px; float:left; text-align: center; font-size: 20px;\n");
+				buf.append("                 font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">" + s + "</div>\n");
+				total += sacks[p][r];
+			}
+			buf.append("    </div>\n");
+			// total marbles
+			buf.append("    <div style=\"width: 54px; height: 44px; float:left; border: 3px solid black; text-align: center;\n");
+			buf.append("                font-size: 20px; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">" + total + "</div>\n");
+			buf.append("    <div style=\"clear:both;\"></div>\n");
+		}
+		// close result array
+		buf.append("   </div>\n");
+		// space between buttons and array
+		buf.append("   <div style=\"width: 400px; height: 100px; float:left;\"></div>\n");
+		buf.append("   <div style=\"clear:both;\"></div>\n");
+		// button 1
+		buf.append("   <div style=\"width: 400px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;\n");
+		buf.append("               font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\"><a href=\"play\">Play</a></div>\n");
+		buf.append("   <div style=\"clear:both;\"></div>\n");
+		// button 2
+		buf.append("   <div style=\"width: 400px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;\n");
+		buf.append("               font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\"><a href=\"stop\">Stop</a></div>\n");
+		buf.append("   <div style=\"clear:both;\"></div>\n");
+		// button 3
+		buf.append("   <div style=\"width: 400px; height: 70px; float:left; cursor: pointer; text-align: center; font-size: 40px;\n");
+		buf.append("               font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\"><a href=\"step\">Step</a></div>\n");
+		buf.append("   <div style=\"clear:both;\"></div>\n");
+		// close left part of page
+		buf.append("  </div>\n");
+		// hexagon
+		buf.append("  <div style=\"width:" + pixels + "px; float: left;\">\n");
 		int dim2_1 = board.length;
 		int dim = (dim2_1 + 1) >> 1;
 		int b = (pixels / dim2_1) - 6;
@@ -286,26 +357,9 @@ public class Cell {
 			buf.append("   </div>\n");
 			buf.append("   <div style=\"clear:both;\"></div>\n");
 		}
+		// close right side of page
 		buf.append("  </div>\n");
-		buf.append("  <div style=\"width: 350px; float: left;\">\n");
-		buf.append("   <div style=\"width: 350px; height: 50px; float:left;\"></div>");
-		buf.append("   <div style=\"clear:both;\"></div>\n");
-		buf.append("   <div style=\"width: 50px; height: 50px; float:left;\"></div>");
-		for (String color : colors)
-			buf.append("   <div style=\"width: 44px; height: 44px; float:left; border: 3px solid black; background-color: " + color + "\"></div>");
-		buf.append("   <div style=\"clear:both;\"></div>\n");
-		for (int p = 0 ; p != players.length ; ++p) {
-			buf.append("   <div style=\"width: 44px; height: 44px; float:left; border: 3px solid black; text-align: center;");
-			buf.append("               font-size: 25px; font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">" + (p + 1) + "</div>\n");
-			buf.append("   <div style=\"width: 294px; height: 44px; float: left; border: 3px solid black;\">\n");
-			for (int r = 0 ; r != 6 ; ++r) {
-				buf.append("    <div style=\"width: 49px; height: 44px; float:left; text-align: center; font-size: 20px;");
-				buf.append("                font-weight: bold; font-family: 'Comic Sans MS', cursive, sans-serif\">" + sacks[p][r] + "</div>\n");
-			}
-			buf.append("   </div>\n");
-			buf.append("   <div style=\"clear:both;\"></div>\n");
-		}
-		buf.append("  </div>\n");
+		// close page
 		buf.append(" </div>\n");
 		buf.append("</body>\n");
 		buf.append("</html>\n");
@@ -317,6 +371,7 @@ public class Cell {
 		System.err.println("---------------");
 		for (int p = 0 ; p != players.length ; ++p) {
 			int[] location = player_location[p];
+			if (location == null) continue;
 			int i = location[0];
 			int j = location[1];
 			System.err.print("Player " + p + ": " + i + "," + j + " [" + sacks[p][0]);
@@ -412,6 +467,7 @@ public class Cell {
 
 	private int[] copyI(int[] a)
 	{
+		if (a == null) return null;
 		int[] b = new int [a.length];
 		for (int i = 0 ; i != a.length ; ++i)
 			b[i] = a[i];
@@ -465,17 +521,33 @@ public class Cell {
 			System.err.println(" (" + possible_moves + ")");
 			if (possible_moves == 0) {
 				in[p] = false;
-				round[p] = -turn;
+				round[p] = -turn - 1;
+				player_location[p] = null;
 				continue;
 			}
 			// move player
-			Player.Direction dir = players[p].move(copyII(board), copyI(player_location[p]),
-			      copyI(sacks[p]), copyII(player_location_copy), copyII(trader_location_copy));
+			Player.Direction dir = null;
+			try {
+				dir = players[p].move(copyII(board), copyI(player_location[p]),
+				                      copyI(sacks[p]), copyII(player_location_copy),
+				                      copyII(trader_location_copy));
+			} catch (Exception e) {
+				System.err.println("Player " + (p + 1) + " threw an exception during move: " + e.getMessage());
+			}
+			if (dir == null) {
+				System.err.println("Invalid move for player " + (p + 1));
+				player_location[p] = null;
+				round[p] = -1;
+				in[p] = false;
+				continue;
+			}
 			System.err.println("Player " + p + " moves " + dir);
 			int[] new_location = move(location, dir);
 			int color = color(new_location, board);
 			if (color < 0 || sack[color] == 0) {
 				System.err.println("Player " + p + ": invalid move");
+				player_location[p] = null;
+				round[p] = -1;
 				in[p] = false;
 				continue;
 			}
@@ -504,7 +576,15 @@ public class Cell {
 				int[] give = new int[6];
 				int p = buyer;
 				int[] sack = sacks[p];
-				players[p].trade(copyD(rates), request, give);
+				try {
+					players[p].trade(copyD(rates), request, give);
+				} catch (Exception e) {
+					System.err.println("Player " + (p + 1) + " threw an exception during trade: " + e.getMessage());
+					player_location[p] = null;
+					round[p] = -1;
+					in[p] = false;
+					continue;
+				}
 				System.err.println("Player " + p + " trades with trader " + t);
 				System.err.print("Rates: [" + rates[0]);
 				for (int r = 1 ; r != 6 ; ++r)
@@ -550,11 +630,14 @@ public class Cell {
 						if (sack[r] < marbles * 4)
 							finished = false;
 					if (finished) {
+						player_location[p] = null;
 						in[p] = false;
 						round[p] = turn;
 					}
 				} else {
 					System.err.println("Player " + p + ": invalid trade");
+					player_location[p] = null;
+					round[p] = -1;
 					in[p] = false;
 				}
 				System.err.println("Transaction is done!");
@@ -629,8 +712,10 @@ public class Cell {
 					System.err.print(r + ": " + players[p].name() + ": ");
 					if (round[p] > 0)
 						System.err.println("finished at round " + round[p]);
-					else if (round[p] < 0)
-						System.err.println("lost marbles at round " + -round[p]);
+					else if (round[p] == -1)
+						System.err.println("disqualified");
+					else if (round[p] < -1)
+						System.err.println("lost marbles at round " + (- round[p] - 1));
 					else
 						System.err.println("remained in game with " + count[p] + " marbles");
 				}
@@ -659,14 +744,14 @@ public class Cell {
 		} else if (dir == Player.Direction.SE) {
 			di = 1;
 			dj = 1;
-		} else
-			throw new Exception("Invalid movement");
+		} else return null;
 		int[] new_location = {i + di, j + dj};
 		return new_location;
 	}
 
 	private static int color(int[] location, int[][] board)
 	{
+		if (location == null) return -1;
 		int i = location[0];
 		int j = location[1];
 		int dim2_1 = board.length;
